init: add ts and basic config to project, add ws server
