+++ .env
@@ -1,0 +1,1 @@
+WS_PORT=3000
